Ex11 complete

+++ Chapter 11 Stenciling/Chapter11_Ex7/BoltApp.cpp
@@ -688,7 +688,7 @@ void BoltApp::BuildDescriptorHeaps()
 	// Create the SRV heap.
 	//
 	D3D12_DESCRIPTOR_HEAP_DESC srvHeapDesc = {};
-	srvHeapDesc.NumDescriptors             = mTextures.size(); //!? look here:  one descriptor for each texture
+	srvHeapDesc.NumDescriptors             = (UINT)mTextures.size(); //!? look here:  one descriptor for each texture
 	srvHeapDesc.Type                       = D3D12_DESCRIPTOR_HEAP_TYPE_CBV_SRV_UAV;
 	srvHeapDesc.Flags                      = D3D12_DESCRIPTOR_HEAP_FLAG_SHADER_VISIBLE;
 	ThrowIfFailed(md3dDevice->CreateDescriptorHeap(&srvHeapDesc, IID_PPV_ARGS(&mSrvDescriptorHeap)));
